Add a golangci-lint files and a custom workflow (#45)
* Add a golangci-lint files and a custom workflow

This add new more strict linter configuration using golangci-lint, this is the same files used accross
mutiples projects. The linter is run using a Github action so we can get the warning directly in the code.
The linter is only done on the new files of this PR, this reduce the run time and allow us to fix incrementally the existing issues.