Add spec for auditbeat
Auditbeat is a wrapped version of Metricbeat and it inherits the same
agent management code as Metricbeat. So the Auditbeat spec resembles
the Metricbeat spec. And this is why Auditbeat doesn't handle the
config from Fleet directly (without the transpiler). Metricbeat would
need to be changed first and Auditbeat would inherit it.

Basically this allows Auditbeat to be invoked by Agent and pass the
input streams through as `auditbeat.module` configuration. Auditbeat
binaries include an `auditbeat.elastic-agent.yml` config file that
Agent will use when starting Auditbeat. This was needed because the
default `auditbeat.yml` has several modules enabled as examples and
we don't want those to run under Agent.