Update containerd to v1.5.10 (#186)
* Update containerd to v1.5.10

Update the containerd module to 1.5.10, version prior to that were
vulnerable.

```
A bug was found in containerd where containers launched through containerd’s CRI implementation with a specially-crafted image configuration could gain access to read-only copies of arbitrary files and directories on the host. This may bypass any policy-based enforcement on container setup (including a Kubernetes Pod Security Policy) and expose potentially sensitive information. Kubernetes and crictl can both be configured to use containerd’s CRI implementation.
```

After looking at our usage of container we are not using this library
to create any containers so we should not be vulnerable to this issue.