Support scheduled actions and cancellation (#419)
* Support scheduled actions and cancellation

Support scheduled actions by adding a new queue that actions will be
added to/removed from before they are sent to the dispatcher. The queue
is a priority queue (ordered by start_time). fleet_gateway is
responsible for syncing the queue to storage. Cancellation of an action
will be handled by a new action dispatcher that will remove actions from
the queue (if any) and update the targetID action status.

TODO
- cancel handler
- action expiration
- fleet_gateway tests

* Add queue tests in fleet_gateway_tests, fix check and linting issues

* Force start_time/expiration to be utc

* Remove logic todos, fix logging statement

* Apply suggestions from code review

Co-authored-by: Anderson Queiroz <me@andersonq.me>

Co-authored-by: Anderson Queiroz <me@andersonq.me>