status identifies failing component, fleet gateway may report degraded, liveness endpoint added (#569)
* Add liveness endpoint

Add /liveness route to metrics server. This route will report the status
from pkg/core/status. fleet-gateway will now report a degraded state if
a checkin fails. This may not propogate to fleet-server as a failed
checkin means communications between the agent and the server are not
working. It may also lead to the server reporting degraded for up to 30s
(fleet-server polling time) when teh agent is able to successfully
connect.

* linter fix

* add nolint direcrtive

* Linter fix

* Review feedback, add doc strings

* Rename noop controller file to _test file