Remove mage notice in favour of make notice
The current implementation of mage notice is not working because it
was never finalised, the fact that it and `make notice` exist only
generates confusion.

This commit removes the `mage notice` and documents that `make notice`
should be used insted for the time being.

In the long run we want to use the implementation on
`elastic-agent-libs`, however it is not woking at the monent.

Closes #1107