[v2] Move queue management to dispatcher (#1109)
* Move queue management to dispatcher

Move queue management actions to the dispatcher from the fleet-server
in order to help with future work to add a retry mechanism. Add a
PersistedQueue type which wrap the ActionQueue to make persisting the
queue simpler for the consumer.

* Refactor ActionQueue

Refactor ActionQueue to only export methods that are used by consumers.
The priority queue implementation has been changed to an unexported
type. Persistency has been added and the persistedqueue type has been
removed.

* Rename persistedQueue interface to priorityQueue

* Review feedback

* failing to save queue will log message

* Chagne gateway to use copy