Add upgrade action retry
Add the ability for the agent to schedule and retry upgrade actions. The
fleetapi actions now define a ScheduledAction, and RetryableAction
interface to eliminate the need for stub methods on all different action
types. Action queue has been changed to function on scheduled actions.
Serialization tests now ensure that that the retry attribute needed by
retryable actions works. A new agent error type (TypeRetryableAction)
has been added to signal when an action can be retried. The handler must
return this type of action to the dispatcher. If the error is returned
and the action is a retryable action, the dispatcher will attempt to
schedule a retry. The dispather will also ack the action to fleet-server
and indicate if it will be retried or has failed (or has been recieved
normally).