Expand status reporter/controller interfaces to allow local reporters (#1285)
* Expand status reporter/controller interfaces to allow local reporters

Add a local reporter map to the status controller. These reporters are
not used when updating status with fleet-server, they are only used to
gather local state information - specifically if the agent is degraded
because checkin with fleet-server has failed. This bypasses the bug that
was introduced with the liveness endpoint where the agent could checkin
(to fleet-server) with a degraded status because a previous checkin
failed. Local reporters are used to generate a separate status. This
status is used in the liveness endpoint.

* fix linter

(cherry picked from commit 717708a72c0c81aded34e93e61b77c9b0cf4e16a)

# Conflicts:
#	internal/pkg/core/status/reporter.go