Use error.TypeLocal to report local issue for liveness endpoint
Add a new ErrorType TypeLocal that can be used to speicify if an error
should effect the state that it reported to fleet-server or if it just
appears under local checks. The current use of this type is to allow a
checkin failure to degraded the state for the liveness endpoint so that
liveness will return a 503 if there is an error, but a subsequent
successful checkin will not report a degraded/failed state.