Remove retryable error type, add action queue CancelType func
Remove the RetryableAction error type. Now a RetryableAction will always
be scheduled for a retry if it returns an error. Acking has changed to
use the Attempt number as an indicator for if an action is being retried
or has failed. If an action has a non-nil error and an attempt count
greater than 0 it will be acked as retried, if it has a non-nil error
and an attempt count less than 1 will be acked as failed. Also made the
action queue cancel upgrade actions if a new upgrade action is retrieved
from the gateway in an attempt to have a single upgrade action (the
latest) apply to an agent.