Add ActionsError() method to config manager
Add a new ActionsError() methdo the the config managers. For the
non-managed instances it will return a nil channel. For the managed
instances it will return the dispatcher error queue directly. Have teh
coordinator gather from this channel as it does for the others and
treat any errors as non-local.