Add diagnostics action to agent
Add a diagnostics action handler to the agent. This will allow a user to
collect a diagnostics bundle from the fleet-ui. Change log collection
from being done in the cmd/diagnostics.go file to dynamically generated
hooks. Change the control server from accepting a list of hooks to a
varadic list of callback functions that return hooks. This is to allow
the agent to dynamically generate a hook for each log (and service log
file) when the diagnostics action is triggered from an action or the
command line. Add a ZipArchive function to the diagnostics package that
will write the zip to the passed writer so that the archive can be
created by a command line invokation or from an action.