Change the stater to include a local flag. (#1308)
* Change the stater to include a local flag.

Change the state reporter to use a local flag that determines if local
errors are included in the resulting state. Assume that configMgr errors
are all local - this effects mainly the fleet_gateway. Allow the gateway
to report an error if a checkin fails. When a checkin fails the local
state reported through the status command and liveness endpoint will
include the error, but checkins to fleet-server will not.

* Add ActionsError() method to config manager

Add a new ActionsError() methdo the the config managers. For the
non-managed instances it will return a nil channel. For the managed
instances it will return the dispatcher error queue directly. Have teh
coordinator gather from this channel as it does for the others and
treat any errors as non-local.

* Fix linter