Diagnostic hooks return times, add start of retrier for uploads
Change the Hook function from returning a []byte to a []byte, time.Time
so that the file mod times collected through diagnostics hooks can be
returned from the same filesystem interaction. Add the start of a
specific retrier for the fleetapi/uploader that will resend requests if
a 429 is recieved.