Add agent diagnostics action
Add the ability for the agent to recieve a diagnostics action. This
action makes the agent take a diagnostics bundle and upload it to a new
fleet-server file upload API. Changed internal details of how
diagnostics bundles are taken to better support this:

Hooks are provided through a callback func so that log files generated
as the agent is running can be included in the bundle. Change the Hook
func signature from returning a []byte to a []byte, time.Time so that
the file mod times collected through diagnostics hooks can be returned
from the same filesystem interaction. Upload client has a built-in
retry mechanism that will resend requests if a 429 is recieved.