Inject input.policy.revision without overwriting.
Ensure each input configuration gets a policy.revision key that contains
the revision number of the Fleet policy, but correctly handle the case
where a policy object already exists.

The policy.revision key was added as a requirement to support endpoint
security, and is probably generally useful for detecting when the
overall agent policy has changed.