Handle requests for /processes/apm-server. (#1792)
Handle requests for /processes/apm-server.

Add special handling for the apm-server process ID for the
/processes/$ID route as well as the /processes response body.

V2 changed the process IDs used for these APIs, which has broken health
checks in the cloud provisioning infrastructure. It is much faster to
add a few lines of code to special case this in the agent than it is to
update all of the cloud monitoring infrastructure.

(cherry picked from commit 378a412b21457f4c383264cc00d7ef4838b6125e)