Remove the changelog GitHub Action check.
It encourages us to create changelog entries that don't need to exist,
it is easier to enforce changelog fragment creation manually in reviews
as we do in Beats.

I will eventually edit this action to enforce that changelog fragments
use valid syntax, but until I do that it is easier to just remove it.