Filter binaries by platform when EXTERNAL=true (#1908)
* Filter binaries by platform when EXTERNAL=true

Not all binaries are available for every platform, and if we try to
download a binary from an artifact URL that doesn't exist the build
fails.

* Update e2e-testing/downloads, handle download errs.

Rather than panicing when we attempt to download an artifact on an
unsupported platform, we ignore the error and log that we are skipping
the download.

* Remove beats dependency.

* Remove redundant word from log

Co-authored-by: Anderson Queiroz <me@andersonq.me>

Co-authored-by: Anderson Queiroz <me@andersonq.me>
(cherry picked from commit a0e6108e4690a25cbc5e80d99d0a8a8c67ebad4c)