Fix sending new processes stale expected states.
When a previously running process exits, it is possible that the most
recent checkin expected message for that process was queued to be sent
in the communicator checkin expected channel but never sent.

When the process is restarted, this stale previous checkin expected
message  would then be sent to the new process as the first checkin.

This can lead to invalid or unexpected initial checkin expected messages
being sent to component processes, for example the configuration index
can be set to the correct value but the actual configuration set to nil
because the stale message was for a process that was already up to date.

Component processes that don't expect this invalid state can then fail
at startup, see https://github.com/elastic/beats/issues/34137 for
example.

To fix this the agent now clears out any pending checkin expected
messages before starting new processes or services.