Fix TestAddToBucket flakiness on Windows
Some testcases insert 3 or more events in a loop to test the rate
limiter.
On windows machines system clock ticks every 15ms so the loop we use to add
events to the rate limiter run slower on those machines (we set the
timer to 2 ms which translates to ~20 ms between iterations on windows).
This commit extends the timeframe of the rate limiter from 50ms to 150ms
in such testcases to make sure that all the events we use in the unit
tests are kept into account.

Closes issue #2146