internal/pkg/diagnostics - Fix creation of sub-directories in logs/ (#2523)
* internal/pkg/diagnostics - Fix creation of sub-directories in logs/

Prior to this change sub-directories within logs/ were not created properly
and resulted in unexpected directories within diagnostics files. A path separator
was missing. Before the change the new test was failing with

    -  Name: (string) (len=13) "logs/sub-dir/",
    +  Name: (string) (len=12) "logssub-dir/",

* Update test to include the versioned sub-directory

I made the test pass, but this these exposes a problem
with the newly added elastic-agent-{commit} directory in
that the zip file is missing a directory entry for it.

* Add missing directory entry for logs/elastic-agent-{commit}

The zip file did not contain a directory entry for the elastic-agent-{commit}
directory.

Relates #2518

* All files within the zip must use forward slash

This calls filepath.ToSlash on any names used in zip entries.

Relates #2518