Add filestream example to the config file
Tailing files is one of the most common use cases but there is no example for it in the config file. This adds the same example as we have in filebeat today: https://github.com/elastic/beats/blob/main/filebeat/filebeat.yml#L21

The change I did compared to the filebeat option is not use `enabled: false`. The reason is that this tripped me in the past. I copy/pased the filebeat example, adjusted it and then was wondering why it didn't work. I didn't read and realise there is `enable: false` in there. So I used comments instead.

In case there is agreement in moving forward with this proposal, the change should also be copied over to elastic-agent.reference.yml