Fix issue with service token path in fleet-server bootstrap of container
Part of the bootstap code in container will read the service token file
in order to make requests against Kibana. When the file is read the
fleet.Elasticsearch.ServiceToken attribute is set and as a result both
--service-token and --service-token-path flags are sent to the enroll
command. This changes the flags to send only --service-token-path if
both attributes are defined.