Add a new job for Sonar only on push on main
This commit adds a new job in the sonarcloud workflow that will be
executed on push events.

Additionally add a push trigger for the main branch, so when new code is
merged on main this will be triggered and send the results into
SonarCloud.

Adding also a workflow_dispatch to be able to trigger it manually

Signed-off-by: Alexandros, Sapranidis <alexandros@elastic.co>