Add ability to filter runner by TEST_PLATFORM (#3181)
Adds a new environment variable that will adjust the integration test
runner to only run tests on the platforms defined in the `TEST_PLATFORM`
variable. This is done in a way that allows the runner to use the
platforms selected to adjust non-specific tests. Meaning if a test is
not specific to the Ubuntu version and the caller has
`TEST_PLATFORM=linux/amd64/ubuntu/20.04` it will use Ubuntu 20.04
instead of the default 22.04.

(cherry picked from commit da5669f081c6842ffe411ee9157bfd6111957cfe)