modify runtime variable 'install.in_default' with rpm/deb (#3309)
* modify runtime variable 'install.in_default' with rpm/deb

modify the `install.in_default` runtime protection variable so that
when elastic-agent is installed via deb or rpm package manager it is
still considered to be in the default path.  This is necessary because
endpoint supports being a component when elastic-agent is installed
via rpm/deb.

* fix typos in comments

(cherry picked from commit 117d7def0e93676f677abfa42e1adb2fd0100e4b)