[Fix] Agent incapable of running on Azure Container Instances (#3576)
What this change is introducing on top of bringing back work introduced in #3084 is change of ordrer for some operations.

Changing owner of a file, discards capabilities set. This becomes a problem with heartbeat as it needs setuid and netraw capabilities to perform properly.
So setting capabilities was moved after chown.