Add support fot fetching artefacts for a build ID
Previously it was not implemented but some of the tests requested a
specific build by its ID. The fetcher was incorrectly returning just the latest
snapshot in this case.

Also, the artifact caching system was not aware of build IDs and
didn't distinguish snapshots of different builds. This led to caching
the latest snapshot and returning it for any requested build ID.