Add support fetching artefacts for a build ID (#4163)
Previously it was not implemented but some of the tests requested a
specific build by its ID. The fetcher was incorrectly returning just the latest
snapshot in this case.

Also, the artifact caching system was not aware of build IDs and
didn't distinguish snapshots of different builds. This led to caching
the latest snapshot and returning it for any requested build ID.

(cherry picked from commit 84e6bc1a64e01d2c52b3d8e87879026151ffa128)

# Conflicts:
#	testing/integration/upgrade_downgrade_test.go