Fix failing upgrade command when gRPC server interrupts connection
The upgrade CLI command creates a client, connects to the Elastic Agent
server and triggers the upgrade process which is supposed to shut down
the agent and start the upgrade. Sometimes, this happens before
the client gets a successful response and the upgrade command returns
an error code although the upgrade succeeded.

This change checks if the connection was interrupted and treats it as
a successful result instead of error.