Fix failing upgrade command when gRPC server interrupts connection (#4519)
The upgrade CLI command creates a client, connects to the Elastic Agent
server and triggers the upgrade process which is supposed to shut down
the agent and start the upgrade. Sometimes, this happens before
the client gets a successful response and the upgrade command returns
an error code although the upgrade succeeded.

This change checks if the connection was interrupted and treats it as
a successful result instead of error.

(cherry picked from commit 4ad7824108ef29ec911efddb7f5f847a8c186ae2)