Do not fail tests when interrupted connection during upgrade
Older agent versions don't properly handle an interrupted connection during
the upgrade. So, we need to handle it in tests.

The check was introduced in 4ad7824108ef29ec911efddb7f5f847a8c186ae2
however it checked a wrong string, it was supposed to check the output
of the command not the actual error string.