Add conn param and conn-skip flag to diagnostics command
Add CONN param to control protocol so the elastic-agent can use CONN as
an additional parameter when requesting diagnostics actions. The CONN
param will be used to execute the optional connection request
diagnostics and included by default. Add a conn-skip flag to the
diagnostics command that will not send the CONN param so request
diagnostics are skipped.