Change approach to execute elastic-agent uninstall command
Change the approach that is taken when "elastic-agent install -f" is ran
to use an exec to run "elastic-agent uninstall -f" in cases where the
agent is installed. this allows the process that runs the uninstall to
use all the correct path values for the installed agent instead of the
values associated with the binary that the install command is ran from.