"install -f" uses exec to uninstall an existing agent (#4965)
* Add explicit check for token and tamper protection in Uninstall func

* fix typo

* Load features from config, fix protection flag load

* Change approach to execute elastic-agent uninstall command

Change the approach that is taken when "elastic-agent install -f" is ran
to use an exec to run "elastic-agent uninstall -f" in cases where the
agent is installed. this allows the process that runs the uninstall to
use all the correct path values for the installed agent instead of the
values associated with the binary that the install command is ran from.

* Add e2e test

* lookup agent binary on path, fix test

* fix test

* Add flag that preserves old approach

* fix typo

* change args format in test

* Use same fixture

* Hide new option

---------

Co-authored-by: Julien Lind <julien.lind@elastic.co>
(cherry picked from commit 1d7b8654ee8f96f92022bce5bd79f75f6823675c)

# Conflicts:
#	internal/pkg/agent/cmd/install.go
#	testing/integration/endpoint_security_test.go