Update event logger configuration via Fleet and environment variable (container command) (#4932)
This commit adds the ability to receive event logger output
configuration via Fleet. Previously only the log level was received
via Fleet and persisted.

Fleet can store the event logger configuration in the overrides
section from the policy, allowing users to change and persist this
configuration. The Elastic-Agent needs this configuration at startup,
so whenever the Elastic-Agent receives a new policy from Fleet, it
compares the event logging output configuration with its current one,
if it is different, it is persisted to disk and the Elastic-Agent
re-execs. When it re-starts it reads the new values from the
persistent store and applies them.

---------

Co-authored-by: Pierre HILBERT <pierre.hilbert@elastic.co>
(cherry picked from commit 72c1ebddf9c04103bff3f79038e601ad7c910975)