Fix TestLogIngestionFleetManaged, TestDebLogIngestFleetManaged (#5375)
* [integration tests] Keep work directory if test fails

This commit enables the work directory used by the integration tests
framework to be kept in the filesystem if the test fails. The full
path of the test directory is printed when the test fails.

* Update pkg/testing/fixture.go

Co-authored-by: Paolo Chilà <paolo.chila@elastic.co>

* Update pkg/testing/fixture.go

Co-authored-by: Paolo Chilà <paolo.chila@elastic.co>

* Remove log in error branch

* [Integration Tests] Generate namespace based on UUIDv4

The namespace generated by the integration tests framework was not
unique among different tests and test runs, so sometimes collisions
would occurs causing some tests to be flaky.

* Add debug logs

* run mage fmt

* Fix TestDebLogIngestFleetManaged

- Remove debug logs
- Make the deb respect the AGENT_KEEP_INSTALLED env var
- Add errors that only happen on deb to allow list

* Improve logs and test error message

---------

Co-authored-by: Paolo Chilà <paolo.chila@elastic.co>
Co-authored-by: Julien Lind <julien.lind@elastic.co>
(cherry picked from commit 66953248f6d548de174162eda1ceb49bee6796ff)