[Integration Test Framework] fix createTempDir
createTempDir register a test cleanup function to remove the folder it
created, however, on Windows, this folder sometimes fails to be
removed because there are still open file handlers for the files
within the folder.

We fix this problem by retrying to remove the folder with a maximum
overall wait time of 3 seconds. This is a very similar approach to
what Go's t.TempDir does.