Fix flaky test Test_runDispatcher/gateway_actions_passed (#5405)
* refctor: make context timeout adjustable per test case

A short context timeout makes some tests flaky, but making it longer across the board would unnecessarily make test runs longer. Let's make it customizable per test case.

* Fix flaky test `Test_runDispatcher/gateway_actions_passed`

The flakiness is a result of a 100 miliseconds not always being enough time to  dispatch the action in the test environment.

The correct way to fix this issue is to run the assertions in an `assert.Eventually` with a longer timeout. Unfortunately, a deficiency in the `testify` testing framework makes it impossible to put the `AssertExpectations` methods in an `assert.Eventually` or `assert.EventuallyWithT` function: https://github.com/stretchr/testify/issues/1414.

The workaround is to increase the context timeout for the `gateway actions passed` test case. I'm increasing it from 100ms to 200ms.

(cherry picked from commit 0efab75f7a891eb3dcd35d9843cc08984cf1e644)