test: do not wait afterRestartDelay in TestCleanup
delay after agent restart is performed to allow agent to tear down all the processes
important mainly for windows, as it prevents removing files which are in use.
In TestCleanup there are no processes and we don't need to wait.
Update Cleanup method to accept delay as an arg and defaulting to afterRestartDelay in prod

TestCleanup speed improves depending on the os:

linux: 8s -> 0.2s
darwin 8s -> 0.2s
windows: 60s -> 0.2s