Rename the JUnit files when running in BK
When running inside BK, we are running the same commands but in
different enviroments, this PR changes this and differentiate each
resulting artifact of JUnit to have a unique name for the step that was
generated.
This in combination with the JUnit annotate plugin provide better
understanding on what failed on which step in the build.

Signed-off-by: Alexandros Sapranidis <alexandros@elastic.co>