Change maintainer label to opencontainers in Dockerfile
The old way of specifying maintainers in a Dockerfile was to use the
MAINTAINER instruction. This is now deprecated and the maintainer label was
used instead.

Nowadays the recommended way to set a maintainer is through the
`org.opencontainers.image.authors` label.

See https://docs.docker.com/reference/build-checks/maintainer-deprecated/.