mage: check that links are live in the OTel README (#5443)
* mage: add check that links are live in the OTel README

This PR adds a utility based on the `link-patrol` command to check if links
in a given file are live. Added a check in the `check` and `otel:readme`
targets to validate all links in the Otel Readme are valid.

* remove unused directive for forbidigo

* update doc for LinkCheck

* update NOTICE.txt

* Update license header with Elastic License 2.0

---------

Co-authored-by: Shaunak Kashyap <ycombinator@gmail.com>
(cherry picked from commit 835cd9c0bffd460f82df444cfeb758e1c4ab4514)

# Conflicts:
#	go.mod
#	go.sum