Use a PodTemplate resource in the Helm Chart
Instead of an untyped .tpl Helm template, use a PodTemplate resource.
This resource is loaded from a file, templated via the tpl function,
converted to a dict to get a specific key, and then converted back to
YAML.

This is a bit convoluted, but it gives us intellisense for the pod
template.