Fix AGENT_COMMIT_OVERRIDE environment variable name
In PR #4403 a new environment variable has been introduced to use a
commit hash consistent with the hash of the elastic-agent-core DRA
binary used. The env var name was not initialized and ended up using
"MANIFEST_URL" which is obviously incorrect.

This commit sets the correct name AGENT_COMMIT_OVERRIDE to avoid using
the manifest url as a commit hash