Emit vars even if provider data is empty from the start (#6598)
We only signal from providers if the emitted data is different from the
current state. This
caused a bug where the data wasn't ever emitted if it was empty. The
provider controller now distinguishes between no data (yet) and empty
data. As a result, it's again possible to effectively disable all
providers by using a local provider with no data.

This is a regression caused by
https://github.com/elastic/elastic-agent/pull/6114. It doesn't apply to
the main and 8.x branches due to
https://github.com/elastic/elastic-agent/pull/6169 refactoring provider
initialisation in those.