ci: remove Dependabot config for OTel components
The pull requests to update OTel components created by Dependabot
have been consistently incomplete and unfit for merging.
They also don't include a changelog entry,
which we want to be included with OTel updates.
We have been relying on manual updates of OTel components for the last couple versions,
with the intention to move to a different automation in the future
that satisfies all requirements (see https://github.com/elastic/elastic-agent/issues/6132).
At this point, the Dependabot PRs are just noise,
so let's remove this automation to get rid of the noise.

Part of https://github.com/elastic/elastic-agent/issues/6132