Add ability to enroll with defined ID and replace_token (#6498)
Allows an Elastic Agent to enroll with a defined ID and replacement token to allow it to replace an existing Elastic Agent. The original Elastic Agent must have also been enrolled with the same --replace-token or it will not be allow to enroll if the --id collides with an existing Elastic Agent.

(cherry picked from commit 8a878fc75ddc6530a42921d6c5cae1fee8740510)