Collapse pipelines under elastic-agent
In order for the Buildkite PR bot to be able to publish meaningful
comments for failures, the pipelines need to be collapsed / uploaded
rather than triggered, otherwise we lose the context and the ability
to have the comment include specific failures for the failed steps in
triggered pipelines.

This commit switches away from triggering to pipeline upload commands
to enable this functionality.