Allow PR comments for failures from the elastic-agent pipeline (#6980)
This comment enables the native PR comment functionality from the Buildkite build bot which will allow comments in the PR about steps that failed even when they are flaky. It is enabled on the parent `elastic-agent` pipeline.

It improves visibility for PR authors and the team about test failures and flakiness.