Ensure permissions are correctly set when extracting archive
On Windows, there was a rare chance that when extracting the archive file
during an upgrade, the permissions were not correctly set leading
OSQuery not to run.

This commit fixes it by enforcing the correct permissions are always
set to every folder and file extracted from the archive for all
platforms.