Fold triggered pipelines under elastic-agent (#6978)
In order for the Buildkite PR bot to be able to publish meaningful comments for failures, the pipelines need to be collapsed / uploaded rather than triggered, otherwise we lose the context and the ability to have the comment include specific failures for the failed steps in triggered pipelines.

This commit switches away from triggering to pipeline upload commands to enable this functionality.

(cherry picked from commit 33d01371e85ed89577b7b1b617fb3fa4cd89186e)

# Conflicts:
#	.buildkite/bk.integration.pipeline.yml