Increase timeout for assertions in TestSetLogLevelFleetManaged (#7190)
Fleet checkins may take up to 5 minutes so we need to increase timeouts
for assertions in TestSetLogLevelFleetManaged to account for it as the
agent may apply log level settings while there's already a Fleet
check-in in progress.
Settings timeouts for retrieving the correct log level in Fleet to 6
minutes should give enough time to agent to set the log level and
communicate it to Fleet in a new check-in request.

(cherry picked from commit 12fa5571af13be8c419a6401645867af518671c2)

# Conflicts:
#	testing/integration/log_level_test.go