Correctly handle sending signal to child process (#7738)
Calling proc.Kill causes the process to exit immediately, which
prevents components from executing some clean up functions.

To gracefully terminate the components, on Linux and MacOS SIGTERM
is sent.

On Windows signals are sent to process groups, child process are created
in the same group as the parent, so sending a signal to a child process
causes the Elastic-Agent to receive the signal as well. Therefore child
process need to be created in their own process group.

To do so, the child process needs to be created setting the creation flag
CREATE_NEW_PROCESS_GROUP, which has the side effect of disbling
the CTRL+C signal.

Then to correctly terminate the child process `windows.GenerateConsoleCtrlEvent`
is called to send CTRL+BREAK signal that is translated by go Go runtime into a
`syscall.SIGINT`, which is already correctly handled by Beats.