fix(diagnostics): handle log rotation races
Handle a race condition that can occur in Agent diagnostics if log
rotation happens while logs are being zipped. There is a time window
between when `filepath.WalkDir` reads the directory contents and when
log files are opened for reading. During this time window, log rotation
can happen, resulting in files that no longer exist and cannot be added
to the diagnostic archive.

To handle this, `fs.ErrNotExist` errors are ignored when attempting to
add log files.