feat: generate a subject key identifier when creating a certificate (#6379)
* feat: generate a subject key identifier when creating a certificate

If a subject key id is omitted, go will generate one using sha1.
This is described as method 1 in RFC 5280 Section 4.2.1.2.

When sha1 is not available (e.g. fips only mode) this method will
panic.

Update the code to explicitly pass a subject key id to avoid calling
sha1 functions. The new SubjectKeyId is generated using
method 1 in RFC 7093 Section 2 which takes 160-bits of the SHA-256 hash.

* Update ca.go

---------

Co-authored-by: Michal Pristas <michal.pristas@gmail.com>