ci: automatically update OTel components (#8288)
* automatically update OTel components

Creates a new GitHub Actions workflow that runs Updatecli daily to check if a new version of OTel components is available. If it is, Updatecli runs the `update-otel.sh` script and creates a new PR with the changes.

These automated updates are currently configured to run for the `main` branch and for the `8.19` branch.

* restore `go mod tidy` in `update-otel.sh`

Forward-porting https://github.com/elastic/elastic-agent/pull/8326

* fix: install changelog tool if missing

* fix: add missing env vars to Updatecli target config

* define branch matrix as a static list

* run Updatecli script from the main branch