Preserve upgrade marker when rolling back and add rollback reason to upgrade details (#8407)
A new reason string field has been added in UpgradeDetails struct. Such field should contain a human-readable description of the cause of the last state transition.
In the specific case this is used to distinguish how elastic-agent got to the UPG_ROLLBACK state: the only possible reason at the moment is an automatic rollback initiated by the elastic-agent watcher, but this could be used for any state transition

This PR will allow to to distinguish if the cause of an UPG_ROLLBACK status is:
-   the new agent misbehaving (automatic rollback)
-    a manual rollback (will be introduced with "Handle rollback command/action in upgrade flow" #6889)


* Add reason for rolled back upgrades

* Don't immediately remove update marker in case of rollback

* Extract repackaging utility functions

* Fix assertion about upgrade marker surviving rollback

* Watcher should recognize terminal upgrade states

* test watch command

Add some unit tests for watchCmd function, covering the most common
cases. This has required some refactoring to properly assert side
effects and interactions.

* Add managed agent rollback integration test