Use a custom image for the ECH deployment when running integration tests (#8337)
Use the image produced in the packaging step as the integrations server
image on ECH when a cloud deployment is provisioned for integrations
tests. This allows for the changes made in a PR to be tested on the
cloud deployment in addition to the current client-focused tests.

(cherry picked from commit c1136e4ab6eff36485359b5fe58e7adb13965c77)

# Conflicts:
#	magefile.go