Update code ownership of core-components.yml
With increased components' movement, the role of `core-components.yml` is becoming critical. The ingest docs team is not qualified to decide on the content of that file.

This PR suggests adding the PM and EM for these file decisions as exclusive owners of the file. That also means they would be gating components updates and changes in each Collector release.