Update code ownership of core-components.yml (#10103)
* Update code ownership of core-components.yml

With increased components' movement, the role of `core-components.yml` is becoming critical. The ingest docs team is not qualified to decide on the content of that file.

This PR suggests adding the PM and EM for these file decisions as exclusive owners of the file. That also means they would be gating components updates and changes in each Collector release.

* Add group

* Update CODEOWNERS

(cherry picked from commit 8079da94fdfe87f4567dbcc57bf13aca6da94250)

# Conflicts:
#	.github/CODEOWNERS