Fix managed rollback test (#11218)
* Add logging to stop/start operation for restartAgentNTimes in Test*UpgradeRollbackOnRestarts

* Do not reset lastPid at every iteration

* Switch to restarting based on version in TestFleetManagedUpgradeRollbackOnRestarts

Move from a fixed number of restarts to restarting only if the agent
matches the target version in the managed upgrade rollback integration
test.
This is done to avoid +-1 errors restarting the agent to simulate only
the upgrade agent crashing, while we can keep restarting for a given
time (controlled by a context with timeout) if the agent matches a
particular version. This should make the test more reliable and avoid
weird interaction between the restart timing and the watcher polling.