Add packaging checks to validate component existence
Add a check to each packaged artifact type to validate that all the expected
components, and no extras, exist in the artifact for the package variant
which is being built.

FIPS packages are not included in the new checks, as they already have a
more sophisticated check in checkFIPS, and a different list of
components.