Use separate docker component list in packaging.
Add a new component list "elastic_agent_docker_components" which is used
only in the container image definition.

This is needed, because despite being defined to have "pkg-type-docker"
only, cloud-defend was being added to linux tar.gz packages, since it
was in the "elastic_agent_components" list.

Now cloud-defend is only in the component list which is used by
agent_docker_spec.