Include only required components for each package type
In the mage packaging scripts, add a filter that will include only
components which support at least one package type.

Previously components would be included even if they did not support any
of the package types being built. cloud-defend is the first component
which does not use pkg-type-all, so it's the first time this problem has
surfaced.