Add cloud-defend to agent container images (#11450)
Restore cloud-defend to the basic and complete Docker images.

This fixes the mage packaging scripts to properly exclude components which don't apply to the package type. Cloud-defend is the first component to not apply to all types, so it revealed some cases where packaging tried to include components to types which they don't apply